FAI-1946 - Update index to show saved searches

+++ src/SFA.DAS.FAA.Domain/SavedSearches/GetSavedSearchesApiResponse.cs
@@ -19,7 +19,7 @@ public record SavedSearchDto(
 public record SearchParametersDto(
     string? SearchTerm,
     List<int>? SelectedRouteIds,
-    decimal? Distance,
+    int? Distance,
     bool DisabilityConfident,
     List<int>? SelectedLevelIds,
     string? Location
